IA_api/requirements.txt
transformers
torch 
fastapi 
uvicorn
lime
shap
google-generativeai
dotenv
